attached the script for counting pickups to the pickups
