+++ Assets/Resources/Prefabs/Pick Ups/Coin.prefab
@@ -23,6 +23,7 @@ GameObject:
   - component: {fileID: 136531832637409480}
   - component: {fileID: 23437683050490066}
   - component: {fileID: 54592767899885394}
+  - component: {fileID: 114722951144958862}
   m_Layer: 0
   m_Name: Coin
   m_TagString: Pick Up
@@ -97,6 +98,17 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &114722951144958862
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1717017528972828}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c16f2f2493a9b034aae23b0e497ca764, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!136 &136531832637409480
 CapsuleCollider:
   m_ObjectHideFlags: 1
@@ -104,7 +116,7 @@ CapsuleCollider:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1717017528972828}
   m_Material: {fileID: 0}
-  m_IsTrigger: 1
+  m_IsTrigger: 0
   m_Enabled: 1
   m_Radius: 0.5
   m_Height: 2
