+++ Assets/Scripts/ScrapValueCoin.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c16f2f2493a9b034aae23b0e497ca764
+timeCreated: 1506474332
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
